+++ plugins/NpgsqlRest.TsClient/TsClientOptions.cs
@@ -6,7 +6,8 @@ public class TsClientOptions(
     bool includeHost = false,
     string? customHost = null,
     CommentHeader commentHeader = CommentHeader.Simple,
-    bool includeStatusCode = false)
+    bool includeStatusCode = false,
+    bool bySchema = false)
 {
     /// <summary>
     /// File path for the generated code. Set to null to skip the code generation.
@@ -35,4 +36,9 @@ public class TsClientOptions(
     /// Set to true to include status code in response: {status: response.status, response: model}
     /// </summary>
     public bool IncludeStatusCode { get; set; } = includeStatusCode;
+
+    /// <summary>
+    /// Create files by PostgreSQL schema. File name will use formatted FilePath where {0} is is the schema name in the pascal case.
+    /// </summary>
+    public bool BySchema { get; set; } = bySchema;
 }